Definida a saida de dados

+++ TaxesPaidConsole/Entities/CompanyPayer.cs
@@ -6,13 +6,13 @@ namespace TaxesPaidConsole.Entities
 {
     class CompanyPayer : Payer
     {
-        public double EmployeesNumber { get; set; }
+        public int EmployeesNumber { get; set; }
 
         public CompanyPayer()
         {
         }
 
-        public CompanyPayer(string name, double anualIncome, double employeesNumber) : base(name, anualIncome)
+        public CompanyPayer(string name, double anualIncome, int employeesNumber) : base(name, anualIncome)
         {
             EmployeesNumber = employeesNumber;
         }
